review endpoints fix

+++ src/main/java/dto/FreelancerDetailsDto.java
@@ -1,5 +1,7 @@
 package dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,7 +9,7 @@ public class FreelancerDetailsDto {
     private String id;
     private String firstName;
     private String lastName;
-    private double rating;
+    private BigDecimal rating;
     private long phoneNumber;
     private String city;
 }
